Modified Function preventNegativeSign
When the "-", negative sign, was prevented, it was also preventing "e" which is used for scientific notation input such as 1e+10, so another condition was added to the if statement's test within it.

Now it is allowing e to be inputed, but a side effect of using input type = number and possibly this modification seems to be that it allows the user to input values such as 3-3 or 3--e and etc, but after checking it out in the console, the value attribute is "", an empty string for such cases.