Add Function numberInputsHaveNumbers
This function checks to see if every input's, type number, is not strictly equal to an empty string, value !== '', which is the default value for the input type number and which is also the value when an invalid number is inputed by the user.

If it returns true, that means that every input has a valid number.