Add CSS Rule Set To Offset Recently Linked Stylesheet Side Effects
Modified File: mocha-in-browser-automated-testing.html

After running the mocha automated tests, it became apparrent that the recently linked stylesheet (multiple-random-number-generator.css) had slightly modified the mocha output elements' css such as centering all the text in the output.

So the #mocha, #mocha * {...} rule set was added to align the text to the left.

The other style sheet side effects that were noticed seem to improve the look of the mocha automated tests' output. So for now, they'll remain as is.