Add Describe Block "function numberInputsHaveNumbers"
As of now, this describe block as two it blocks. The 1st it block test to see if every input, type number, has a value attribute that is not empty which means they have a valid number, and the other block test if every input's value attribute is empty which means they do not have a valid number.

modified:   mocha-in-browser-automated-tests.js