Add Describe Block "function randomNumberGenerator - floating point number output"
This describe block currently contains a function that creates an it block test that checks to see if the return value of randomNumberGenerator is a floating point number when mrngOutputIntegers' checked attribute is false, and the mrngMinNumber.value and mrngMaxNumber.value are equal to a floating point number.

A while loop continues to invoke the it block making function until it's stop, break condition is reached.

modified:   mocha-in-browser-automated-tests.js