Add Function makeWithinMinMaxTest and While Loop
This function makes an it block that test if the value returned by randomNumberGenerator is within the min and max values, numbers.

The it block contains several if statements that check if the returned number, value is a whole number, > max, or < min and if it meets any of those conditions, a failure message will be shown.

The while loop runs the function until its condition becomes false.

modified:   mocha-in-browser-automated-tests.js