Create Describe Block "function numbersToGenerate"
This describe block will be used to verify if numbersToGenerate is appending the appropriate number of numbers or nothing to mrngOutput, paragraph element, based on input elements' attribute values that are set, every time it is run.

The initial it block in it checks to see if the mrngOutput.innerText is equal to "" when mrngNumberOfRuns.value = 0.

modified:   mocha-in-browser-automated-tests.js