Add Function returnNumber to D. Block "function numbersToGenerate"
This function will return a negative or positve whole number or floating point number based on the the three arguments it receives (number, type, sign) which have default values set just in case no parameters are given to it.

This function can be used to randomly generate a number for some parameters in the makeNumbersToGenerateTest function.

modified:   mocha-in-browser-automated-tests.js