Add Function Expression, Variables and While Loop
The function expression randomNumber was added to be used in the condition of the question mark operator which returns true or false for the allowNegatives variable.

Variable startCount, callCount, breakLoop were added.

The while loop will continue to invoke makeNumbersToGenerateTest until the callCount is less than or equal to the breakLoop variable.

modified:   mocha-in-browser-automated-tests.